@@ -1,5 +1,5 @@
 import { ElementRef, EventEmitter, OnDestroy, OnChanges, SimpleChanges } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 /**
  * Draggable Directive for Angular2
  *
