@@ -10,9 +10,9 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 var core_1 = require("@angular/core");
+var rxjs_1 = require("rxjs");
 var events_1 = require("../events");
 var operators_1 = require("rxjs/operators");
-var fromEvent_1 = require("rxjs/observable/fromEvent");
 var ResizeableDirective = /** @class */ (function () {
     function ResizeableDirective(element, renderer) {
         this.renderer = renderer;
@@ -50,10 +50,10 @@ var ResizeableDirective = /** @class */ (function () {
         if (isHandle) {
             event.stopPropagation();
             this.resizing = true;
-            var mouseup = fromEvent_1.fromEvent(document, 'mouseup');
+            var mouseup = rxjs_1.fromEvent(document, 'mouseup');
             this.subscription = mouseup
                 .subscribe(function (ev) { return _this.onMouseup(); });
-            var mouseMoveSub = fromEvent_1.fromEvent(document, 'mousemove')
+            var mouseMoveSub = rxjs_1.fromEvent(document, 'mousemove')
                 .pipe(operators_1.takeUntil(mouseup))
                 .subscribe(function (e) { return _this.move(e, initialWidth, mouseDownScreenX); });
             this.subscription.add(mouseMoveSub);
