@@ -3,9 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var camel_case_1 = require("./camel-case");
 var cache = {};
 var testStyle = typeof document !== 'undefined' ? document.createElement('div').style : undefined;
-// Get Prefix
-// http://davidwalsh.name/vendor-prefix
-var prefix = function () {
+var ɵ0 = function () {
     var styles = typeof window !== 'undefined' ? window.getComputedStyle(document.documentElement, '') : undefined;
     var match = typeof styles !== 'undefined' ?
         Array.prototype.slice.call(styles).join('').match(/-(moz|webkit|ms)-/) : null;
@@ -17,7 +15,11 @@ var prefix = function () {
         css: "-" + pre + "-",
         js: pre[0].toUpperCase() + pre.substr(1)
     } : undefined;
-}();
+};
+exports.ɵ0 = ɵ0;
+// Get Prefix
+// http://davidwalsh.name/vendor-prefix
+var prefix = ɵ0();
 function getVendorPrefixedName(property) {
     var name = camel_case_1.camelCase(property);
     if (!cache[name]) {
