@@ -6,6 +6,7 @@ import {
   DataTableRowWrapperComponent,
   DataTableBodyCellComponent,
   DataTableSelectionComponent,
+  DataTableSummaryRowComponent,
   ProgressBarComponent,
   ScrollerComponent
 } from '.';
@@ -19,12 +20,13 @@ describe('DataTableBodyComponent', () => {
   // provide our implementations or mocks to the dependency injector
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [ 
+      declarations: [
         DataTableBodyComponent,
         DataTableBodyRowComponent,
         DataTableRowWrapperComponent,
         DataTableBodyCellComponent,
         DataTableSelectionComponent,
+        DataTableSummaryRowComponent,
         ProgressBarComponent,
         ScrollerComponent
       ],
@@ -37,7 +39,7 @@ describe('DataTableBodyComponent', () => {
       fixture = TestBed.createComponent(DataTableBodyComponent);
       component = fixture.componentInstance;
       element = fixture.nativeElement;
-    });    
+    });
   }));
 
   describe('fixture', () => {
@@ -45,7 +47,6 @@ describe('DataTableBodyComponent', () => {
       expect(component).toBeTruthy();
     });
   });
-
 
   describe('Paging', () => {
 
@@ -55,7 +56,7 @@ describe('DataTableBodyComponent', () => {
       component.pageSize = 10;
       component.offset = 1;
       component.rowCount = 20;
-      let expectedIndexes = { first: 10, last: 20};
+      const expectedIndexes = { first: 10, last: 20};
       component.updateIndexes();
       expect(component.indexes).toEqual(expectedIndexes);
     });
@@ -66,7 +67,7 @@ describe('DataTableBodyComponent', () => {
       component.pageSize = 5;
       component.offset = 1;
       component.rowCount = 9;
-      let expectedIndexes = { first: 5, last: 9};
+      const expectedIndexes = { first: 5, last: 9};
       component.updateIndexes();
       expect(component.indexes).toEqual(expectedIndexes);
     });
@@ -77,7 +78,7 @@ describe('DataTableBodyComponent', () => {
       component.pageSize = 10;
       component.offset = 1;
       component.rowCount = 20;
-      let expectedIndexes = { first: 0, last: 10};
+      const expectedIndexes = { first: 0, last: 10};
       component.updateIndexes();
       expect(component.indexes).toEqual(expectedIndexes);
     });
@@ -88,7 +89,7 @@ describe('DataTableBodyComponent', () => {
       component.pageSize = 5;
       component.offset = 1;
       component.rowCount = 9;
-      let expectedIndexes = { first: 0, last: 5};
+      const expectedIndexes = { first: 0, last: 5};
       component.updateIndexes();
       expect(component.indexes).toEqual(expectedIndexes);
     });
