@@ -1,6 +1,6 @@
 declare let global: any;
 
 /* tslint:disable:variable-name */
-export const MouseEvent = (global as any).MouseEvent as MouseEvent;
-export const KeyboardEvent = (global as any).KeyboardEvent as KeyboardEvent;
-export const Event = (global as any).Event as Event;
+export const MouseEvent = ((window as any) || (global as any)).MouseEvent as MouseEvent;
+export const KeyboardEvent = ((window as any) || (global as any)).KeyboardEvent as KeyboardEvent;
+export const Event = ((window as any) || (global as any)).Event as Event;
